requirements.txt.txt
streamlit>=1.30
tensorflow-cpu>=2.14
numpy>=1.23
pillow>=9.5
scikit-learn>=1.2
